day 17
hard to figure out because the example didnt highlight that the frame of reference was changing to the active nodes so I was going crazy since I can't read. after that pretty easy and my idea almost worked at first I just had to fix the counting because I chose the wrong thing once.

probably one of the easier ones of these to read and comprehend at least on the side of my code. partially cause of the comments partially cause its pretty straight forward just a pain