day 20
i almost dont want to submit this the part two is so gross and just completely bad but I want to for posterity. I got 189/1410 in 18;33/4:41:42 so im in pain

The solution to the first part seemed hard but then I realized I just needed to find the corners so I could do that without builidng the image, then I really struggled figuring how to build the image and basically slapped on more code over and over when I couldve probably gotten it done in half the time if id taken the time to think about it. I just dug myself to deep into the hole to get out. ill fix it tomorrow probly